[Perf] Remove some closures in Stream
Related to #147

Node 10.0.0
Reference = 38321d32586348f4c7d5793a93127f7567c04cdf
Test = this commit

Reference x 1.59 ops/sec ±2.77% (24 runs sampled)
Test x 1.84 ops/sec ±1.83% (24 runs sampled)
Reference x 1.54 ops/sec ±2.48% (23 runs sampled)
Test x 1.70 ops/sec ±2.95% (24 runs sampled)

Reference (target=es6) x 2.16 ops/sec ±1.13% (25 runs sampled)
Test (target=es6) x 6.26 ops/sec ±2.41% (35 runs sampled)
Reference (target=es6) x 2.12 ops/sec ±1.10% (25 runs sampled)
Test (target=es6) x 6.36 ops/sec ±3.81% (35 runs sampled)
